Fixes #1478
