+++ management/server/core/hub-manager/hub-manager-impl/src/main/java/io/subutai/core/hubmanager/impl/LogListenerImpl.java
@@ -1,0 +1,70 @@
+package io.subutai.core.hubmanager.impl;
+
+
+import java.security.MessageDigest;
+import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.Set;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import org.apache.commons.lang.ArrayUtils;
+
+import io.subutai.core.appender.SubutaiErrorEvent;
+import io.subutai.core.appender.SubutaiErrorEventListener;
+
+
+public class LogListenerImpl implements SubutaiErrorEventListener
+{
+    private final Logger log = LoggerFactory.getLogger( getClass() );
+
+    private final Map<String, String> errLogs = new LinkedHashMap<>();
+
+
+    public Set<String> getErrLogs()
+    {
+        Set<String> logs = new HashSet<>();
+
+        synchronized ( errLogs )
+        {
+            logs.addAll( errLogs.values() );
+            errLogs.clear();
+        }
+
+        return logs;
+    }
+
+
+    @Override
+    public void onEvent( final SubutaiErrorEvent event )
+    {
+        log.info( String.format( "RECEIVED:%n:%s", event.toString() ) );
+
+        try
+        {
+            byte[] loggerName = event.getLoggerName().getBytes();
+            byte[] renderedMsg = event.getRenderedMessage().getBytes();
+            byte[] combined = ArrayUtils.addAll( loggerName, renderedMsg );
+            MessageDigest md = MessageDigest.getInstance( "MD5" );
+            byte[] theDigest = md.digest( combined );
+            String key = new String( theDigest );
+
+            synchronized ( errLogs )
+            {
+                errLogs.put( key, event.toString() );
+
+                while ( errLogs.size() > 10 )
+                {
+                    //delete oldest value
+                    errLogs.remove( errLogs.keySet().iterator().next() );
+                }
+            }
+        }
+        catch ( Exception e )
+        {
+            log.warn( "Error in #onEvent {}", e.getMessage() );
+        }
+    }
+}
