+++ management/server/core/hub-manager/hub-manager-impl/src/main/resources/OSGI-INF/blueprint/context.xml
@@ -13,11 +13,11 @@
   <authz:enable />
   <!-- ***************************************************************** -->
 
-  <bean id="errLstnr" scope="singleton" class="io.subutai.core.hubmanager.impl.processor.HubLoggerProcessor" />
-  <service ref="errLstnr" interface="io.subutai.core.appender.SubutaiErrorEventListener" />
-
   <bean id="hostLstnr" scope="singleton" class="io.subutai.core.hubmanager.impl.processor.ResourceHostDataProcessor" />
   <service ref="hostLstnr" interface="io.subutai.core.hostregistry.api.HostListener" />
+
+  <bean id="logListener" scope="singleton" class="io.subutai.core.hubmanager.impl.LogListenerImpl" />
+  <service ref="logListener" interface="io.subutai.core.appender.SubutaiErrorEventListener" />
 
   <reference id="securityManager"
              interface="io.subutai.core.security.api.SecurityManager" />
@@ -55,6 +55,7 @@
     <property name="hostRegistrationManager" ref="hostRegistrationManager" />
     <property name="monitor" ref="monitor" />
     <property name="commandExecutor" ref="commandExecutor" />
+    <property name="logListener" ref="logListener" />
     <argument ref="daoManagerHub" />
   </bean>
